+++ src/AlphaSquad.Api/GlobalUsings.cs
@@ -26,6 +26,7 @@ global using AlphaSquad.Shared.DTOs.Store;
 global using AlphaSquad.Shared.DTOs.Social;
 global using AlphaSquad.Shared.DTOs.Gamification;
 global using AlphaSquad.Shared.DTOs.Events;
+global using AlphaSquad.Shared.DTOs.Legal;
 global using AlphaSquad.Api.Features.Classes;
 global using AlphaSquad.Api.Features.Exercises;
 global using AlphaSquad.Api.Features.Workouts;
@@ -36,6 +37,7 @@ global using AlphaSquad.Api.Features.Store;
 global using AlphaSquad.Api.Features.Social;
 global using AlphaSquad.Api.Features.Gamification;
 global using AlphaSquad.Api.Features.Events;
+global using AlphaSquad.Api.Features.Legal;
 global using AlphaSquad.Infrastructure.Features;
 global using AlphaSquad.Infrastructure.Gamification;
 // System namespaces